+++ docs/Whats-New.md
@@ -375,6 +375,7 @@ New:
 - [Passenger-based insignias](Fixed-or-Improved-Logics.md#customizable-veterancy-insignias) (by Ollerus)
 - [Use `InsigniaType` to set the properties of insignia in a batch](Miscellanous.md#insignia-type) (by Ollerus)
 - Customized Vehicle Turret Rotation (by CrimRecya)
+- Extra threat value coefficient related to orientation (by CrimRecya)
 
 Vanilla fixes:
 - Fixed sidebar not updating queued unit numbers when adding or removing units when the production is on hold (by CrimRecya)